division of 1d and 2d file timeseries
