--- docs/1_geomatics_api_docs/times.rst
@@ -1,19 +1,0 @@
-*****
-times
-*****
-
-Timeseries Functions
-~~~~~~~~~~~~~~~~~~~~
-These functions will generate a timeseries from 1 netcdf file with many timesteps or many netcdf files each containing
-a single timestep.
-
-.. automodule:: geomatics.times
-	:members: point_series, box_series,	shp_series
-
-NCML
-~~~~
-A tool for creating a netCDF Markup Language (NCML) file based on the absolute file paths to datasets.
-
-.. automodule:: geomatics.times
-	:members: gen_ncml
-	:noindex:
