+++ geomatics/__init__.py
@@ -2,4 +2,5 @@ import geomatics._utils
 import geomatics.convert
 import geomatics.data
 import geomatics.inspect
-import geomatics.times
+import geomatics.series_1d
+import geomatics.series_2d
